chore: introduce 'VolatileAggregate' for the aggregate, instead of 'VolatileFragment'
  Currently, both types are identical, but changes will come in the next commit. This current commit simply re-organises the code, modules and methods between those types.
